[skip ci] add MemoryCache trait
