+++ src/Ubiquity/cache/database/HashQueryCache.php
@@ -3,28 +3,23 @@
 namespace Ubiquity\cache\database;
 
 use Ubiquity\cache\system\ArrayCache;
+use Ubiquity\cache\database\traits\MemoryCacheTrait;
 
 /**
  * Ubiquity\cache\database$HashQueryCache
  * This class is part of Ubiquity
  *
  * @author jcheron <myaddressmail@gmail.com>
- * @version 1.0.2
+ * @version 1.0.3
  *
  */
 class HashQueryCache extends DbCache {
+	use MemoryCacheTrait;
 	/**
 	 *
 	 * @var int
 	 */
 	protected $size;
-	/**
-	 *
-	 * @var array
-	 */
-	protected $arrayCache;
-	protected $storeDeferred;
-	protected $toStore = [ ];
 
 	protected function hash(string $string) {
 		return \substr ( \md5 ( $string ), 0, $this->size );
@@ -39,36 +34,19 @@ class HashQueryCache extends DbCache {
 	public function store($tableName, $condition, $result) {
 		$refString = $tableName . $condition;
 		$k = $this->hash ( $refString );
-		$this->getArrayCache ( $k );
-		$this->arrayCache [$k] [$this->getKey ( $refString )] = $result;
+		// $this->getMemoryCache ( $k );
+		$this->memoryCache [$k] [$this->getKey ( $refString )] = $result;
 		if ($this->storeDeferred) {
 			$this->toStore [] = $k;
 		} else {
-			$this->cache->store ( $k, 'return ' . $this->asPhpArray ( $this->arrayCache [$k] ) . ';' );
+			$this->cache->store ( $k, 'return ' . $this->asPhpArray ( $this->memoryCache [$k] ) . ';' );
 		}
-	}
-
-	public function storeDeferred() {
-		foreach ( $this->toStore as $k ) {
-			$this->cache->store ( $k, 'return ' . $this->asPhpArray ( $this->arrayCache [$k] ) . ';' );
-		}
-		$this->toStore = [ ];
-	}
-
-	protected function getArrayCache($key) {
-		if (isset ( $this->arrayCache [$key] )) {
-			return $this->arrayCache [$key];
-		}
-		if ($this->cache->exists ( $key )) {
-			return $this->arrayCache [$key] = $this->cache->fetch ( $key );
-		}
-		return false;
 	}
 
 	public function fetch($tableName, $condition) {
 		$refString = $tableName . $condition;
 		$k = $this->hash ( $refString );
-		if ($cache = $this->getArrayCache ( $k )) {
+		if ($cache = $this->getMemoryCache ( $k )) {
 			$key = $this->getKey ( $refString );
 			if (isset ( $cache [$key] ))
 				return $cache [$key];
@@ -79,7 +57,7 @@ class HashQueryCache extends DbCache {
 	public function delete($tableName, $condition) {
 		$refString = $tableName . $condition;
 		$k = $this->hash ( $refString );
-		if ($cache = $this->getArrayCache ( $k )) {
+		if ($cache = $this->getMemoryCache ( $k )) {
 			$key = $this->getKey ( $refString );
 			if (isset ( $cache [$key] )) {
 				unset ( $cache [$key] );
