+++ src/Ubiquity/cache/database/QueryCache.php
@@ -2,31 +2,48 @@
 
 namespace Ubiquity\cache\database;
 
+use Ubiquity\cache\database\traits\MemoryCacheTrait;
+use Ubiquity\cache\system\ArrayCache;
+
 /**
  * Ubiquity\cache\database$QueryCache
  * This class is part of Ubiquity
  *
  * @author jcheron <myaddressmail@gmail.com>
- * @version 1.0.1
+ * @version 1.0.2
  *
  */
 class QueryCache extends DbCache {
+	use MemoryCacheTrait;
+	protected $storeDeferred;
+
+	public function __construct($cacheSystem = ArrayCache::class, $config = [ ]) {
+		parent::__construct ( $cacheSystem, $config );
+		$this->storeDeferred = $config ['deferred'] ?? true;
+	}
 
 	public function fetch($tableName, $condition) {
-		$key = $tableName . '.' . $this->getKey ( $condition );
-		if ($this->cache->exists ( $key ))
-			return $this->cache->fetch ( $key );
-		return false;
+		return $this->getMemoryCache ( $tableName . '.' . $this->getKey ( $condition ) );
 	}
 
 	public function store($tableName, $condition, $result) {
-		$this->cache->store ( $tableName . '.' . $this->getKey ( $condition ), 'return ' . $this->asPhpArray ( $result ) . ';' );
+		$key = $tableName . '.' . $this->getKey ( $condition );
+		$this->memoryCache [$key] = $result;
+		if ($this->storeDeferred) {
+			$this->toStore [] = $key;
+		} else {
+			$this->cache->store ( $key, 'return ' . $this->asPhpArray ( $result ) . ';' );
+		}
 	}
 
 	public function delete($tableName, $condition) {
 		$key = $tableName . '.' . $this->getKey ( $condition );
-		if ($this->cache->exists ( $key ))
+		if ($this->cache->exists ( $key )) {
+			if (isset ( $this->memoryCache [$key] )) {
+				unset ( $this->memoryCache [$key] );
+			}
 			return $this->cache->remove ( $key );
+		}
 		return false;
 	}
 }
